requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pytz==2023.3
selenium==4.11.2
webdriver-manager==3.8.6
python-telegram-bot[job-queue]==20.3
python-dotenv==1.0.0